% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diffusion_data}
\alias{diffusion_data}
\title{prepared data for the analysis of the diffusion of the new game app among pupils.}
\format{
A data frame (tibble) with eleven variables:
\describe{
  \item{ID}{integer pupil ID, key to other data frames}
  \item{label}{pupil's first name, as a character string}
  \item{sex}{pupil's sex, boy (0) or girl(1)}
  \item{ethnicity}{pupil's ethnicity, coded as a single letter (A, B, C, D, E)}
  \item{adhd}{pupil's ADHD level, coded as a real number between 0 and 10}
  \item{breakID}{integer sequential number of the break in which the statement was made (constant: only Break 1)}
  \item{breakAdopt}{the sequential number of the break in which the pupil has the game app on their device for the first time, as integer}
  \item{adoption}{does the pupil have the game app on their device, as integer: yes (1) or no (0)}
  \item{exposure}{number of minutes that a pupil played with a peer who was playing the new game app in the preceding break, summed over all peers with whom the pupil was playing (at the same time), as real number}
  \item{friends_adopted}{number of a pupil's friends who adopted the new game app in the previous break}
  \item{cum_friends_adopted}{total number of a pupil's friends who adopted the new game app before the current break}
  }
}
\usage{
diffusion_data
}
\description{
For each pupil-break combination (Breaks 1-10) an observation (row) has been
created up to and including the first break in which the pupil has the new
game app on their device.
}
\keyword{datasets}
