% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loudness_average}
\alias{loudness_average}
\title{prepared data for the analysis of statement loudness at the pupil level.}
\format{
A data frame (tibble) with six variables:
\describe{
  \item{ID}{integer pupil ID, key to other data frames}
  \item{avg_loudness}{average loudness of all statements made by a pupil during Break 1, both those directed at another pupil and those that are not; as real number}
  \item{label}{pupil's first name, as a character string}
  \item{sex}{pupil's sex, boy (0) or girl(1)}
  \item{ethnicity}{pupil's ethnicity, coded as a single letter (A, B, C, D, E)}
  \item{adhd}{pupil's ADHD level, coded as a real number between 0 and 10}
  }
}
\usage{
loudness_average
}
\description{
For each pupil, average loudness (during Break 1) has been calculated (from
basic data set pupils_dyn) and personal characteristics have been added (from
basic data set pupils_const).
}
\keyword{datasets}
