% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loudness_events}
\alias{loudness_events}
\title{prepared data for the analysis of statement loudness at the statement level.}
\format{
A data frame (tibble) with fourteen variables:
\describe{
  \item{ID}{integer pupil ID, key to other data frames}
  \item{onset}{time point within the break at which a pupil starts an action, as real number indicating the (fraction of) the minute within the break}
  \item{terminus}{time point within the break at which a pupil ends an action, as real number indicating the (fraction of) the minute within the break}
  \item{loudness}{(average) loudness of pupil's voice, as a real number between 0.05 (whisper) to 1 (loudest cry)}
  \item{expo_last_max}{loudness of the loudest voice at the moment a pupil starts an utterance, as real number}
  \item{expo_last_mean}{he average sound level at the moment a pupil starts an utterance, as real number}
  \item{expo_last_min}{loudness of the softest voice at the moment a pupil starts an utterance, as real number}
  \item{expo_minute_max_playmate}{loudness of the loudest voice of playmates' last utterances in the preceding minute, as real number}
  \item{expo_minute_max_conversation}{loudness of the loudest voice of conversation partners' last utterances in the preceding minute, as real number}
  \item{expo_minute_max_other}{loudness of the loudest voice of other pupils' last utterances in the preceding minute, as real number}
  \item{label}{pupil's first name, as a character string}
  \item{sex}{pupil's sex, boy (0) or girl(1)}
  \item{ethnicity}{pupil's ethnicity, coded as a single letter (A, B, C, D, E)}
  \item{adhd}{pupil's ADHD level, coded as a real number between 0 and 10}
  }
}
\usage{
loudness_events
}
\description{
For each statement during Break 1, loudness is recorded with pupil, start and
end time (from basic data set pupils_dyn). Pupil characteristics have been
added from basic data set pupils_const. Several indicators of exposure to the
loudness of other pupils have been calculated.
}
\keyword{datasets}
