% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NikitaRocks.R
\docType{package}
\name{NikitaRocks}
\alias{NikitaRocks}
\title{NikitaRocks: Tutorials for a summer course on models for events in social networks}
\usage{
learnr::run_tutorial(name, package = "NikitaRocks")

Starts the tutorial named *name*.
}
\description{
The NikitaRocks package provides interactive R tutorials developed for
the St Petersburg Summer School on Network Analysis, July 2021.
}
\examples{

\dontrun{
  learnr::run_tutorial("Session1", package = "NikitaRocks")
 }
}
