% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairs_dyn}
\alias{pairs_dyn}
\title{Time-varying characteristics of the pupils.}
\format{
A data frame (tibble) with nine variables:
\describe{
  \item{from}{integer pupil ID for the sender, key to other data frames}
  \item{to}{integer pupil ID for the receiver, key to other data frames}
  \item{breakID}{integer sequential number of the break in which the statement was made (constant: only Break 1)}
  \item{onset}{time point within the break at which a pupil starts the statement, as real number indicating the (fraction of) the minute within the break}
  \item{onset-censored}{is the start of the tie or interaction observed and initiated by the pupils themselves?; boolean}
  \item{terminus}{time point within the break at which a pupil ends the statement, as real number indicating the (fraction of) the minute within the break}
  \item{terminus-censored}{is the end of the tie or interaction observed and initiated by the pupils themselves?; boolean}
  \item{dyntie}{type of dynamic tie or interaction, coded as text: "Playmate" or "Utterance}
  \item{negative}{the valence of utterances as integer: negative (1) or positive/not negative (0), NA for playmate ties}
}
}
\usage{
pairs_dyn
}
\description{
The data describe ties and interactions among the 26 fictional pupils that
change over time. It contains utterances (directed, one pupil speaking to
another) for Break 1 and playmate ties (undirected, both pupils appear as
sender and as receiver) for Breaks 1 to 10.
This is one of the four basic data frames from which all analysis data frames
have been derived.
The data have been simulated for the purpose of illustrating different
statistical models for events happening within networks.
}
\keyword{datasets}
