% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pupils_const}
\alias{pupils_const}
\title{Time-constant characteristics of the pupils.}
\format{
A data frame (tibble) with eight variables:
\describe{
  \item{ID}{integer pupil ID, key to other data frames}
  \item{label}{pupil's first name, as a caracter string}
  \item{present}{presence of the pupil at the start of data collection, as boolean; required for visualizations with the ndtv package}
  \item{onset}{time point at which the pupil appears within the data set, as integer; required for visualizations with the ndtv package}
  \item{terminus}{time point at which the pupil appears within the data set for the last time, as integer; required for visualizations with the ndtv package}
  \item{sex}{pupil's sex, boy (0) or girl(1)}
  \item{ethnicity}{pupil's ethnicity, coded as a single letter (A, B, C, D, E)}
  \item{adhd}{pupil's ADHD level, coded as a number between 0 and 10}
}
}
\usage{
pupils_const
}
\description{
The data describe characteristics of 26 fictional pupils that do not change.
This is one of the four basic data frames from which all analysis data frames
have been derived.
The data have been simulated for the purpose of illustrating different
statistical models for events happening within networks.
}
\keyword{datasets}
