% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pupils_dyn}
\alias{pupils_dyn}
\title{Time-varying characteristics of the pupils.}
\format{
A data frame (tibble) with eight variables:
\describe{
  \item{ID}{integer pupil ID, key to other data frames}
  \item{breakID}{sequential number of the break, as integer (1 - 10)}
  \item{onset}{time point within the break at which a pupil starts an action, as real number indicating the (fraction of) the minute in the break}
  \item{onset.censored}{indicates whether the start time (onset) was unobserved, as boolean}
  \item{terminus}{time point within the break at which a pupil ends an action, as real number indicating the (fraction of) the minute in the break}
  \item{terminus.censored}{indicates whether the end time (terminus) was unobserved or terminated by circumstances, as boolean}
  \item{loudness}{(average) loudness of pupil's voice, as a real number between 0.05 (whisper) to 1 (loudest cry)}
  \item{gameapp}{pupil is playing a new game app on their own mobile device, as integer: 0 (No) or 1 (Yes)}
}
}
\usage{
pupils_dyn
}
\description{
The data describe characteristics of 26 fictional pupils that may change,
namely the actions of using their voices and playing a new game app.
This is one of the four basic data frames from which all analysis data frames
have been derived.
The data have been simulated for the purpose of illustrating different
statistical models for events happening within networks.
}
\keyword{datasets}
